Move SpaceFillingCanvas into Canvas folder
